major change set (below from email) ---------------------------
Postgres does not allow cross-database queries, something that we use all the time throughout the codebase. Consequently, Paul and Jesse wanted a way to be able to support cross-database like functionality in postgres. They found that postgres supports a notion of "schemas" which are essentially namespaces that group tables -- basically another level of organization under databases. So a database can contain a number of schemas which in turn contain tables. If you have table y in schema x, you can write queries like "select a from x.y". In effect, schemas act _the same way_ as databases do in, say, mysql.

I have modified the database server / opus database framework in order to accommodate this by faking schemas for databases. Specifically, all "create_database" and "drop_database" operations on a _postgres_ connection will actually only create or drop a schema, and all the get_table operations, etc, will look in the correct schema. But the schemas need to live in a database. The new convention is that you will be using a database with the same name as the current project (as set in (1) above). 

The code should take care of this for you. In other words, the database management infrastructure will, by default, create & connect to a database of the same name as the project. You can then create "schemas" inside of this project database using the same interface that you use to create databases for other database engines. For example, in the gui when you execute an indicator batch in the eugene_gridcell project that writes results to a postgres database, and you specify "test_indicators" as the database name, your results will be stored in a table in the "test_indicators" schema of the "eugene_gridcell" database on the postgres server.


git-svn-id: https://svn.urbansim.org/trunk@14227 1513fa5d-83ce-422e-ae47-6967bb0fb594