bayarea: fix hudson's ability to restart runs
The problem is that if we want to restart a run, we must have
access to the run directory.  But because hudson may launch the
model on any one of his slaves, the run directory may not be
available.  The solution is to mount the directory over sshfs.  I
chose this over rsync'ing the run directory because the latter
implies a problematic coherency issue.

git-svn-id: https://svn.urbansim.org/src/trunk@18955 1513fa5d-83ce-422e-ae47-6967bb0fb594