bayarea: create basic django app to serve up opus db
We run the bayarea model on many different model servers in
parallel using hudson.  Hudson offeres a convenient web interface
and REST API for viewing all of the state of all of the runs, but
it has some limitations.  In particular, hudson does not have
access to urbansim internals.  So hudson can't report things
about the run configuration including run numbers and cache
directories.  So we're augmenting the hudson REST API with this
django app.  This alows a client to get run state from the
services DB.  This, in turn, allows us to enrich our mtcstat
front-end with more useful run data.

git-svn-id: https://svn.urbansim.org/src/trunk@19145 1513fa5d-83ce-422e-ae47-6967bb0fb594