bayarea: introducing the hopus tool
hopus stands for hudson + opus.  This is actually a bayarea
specific tool at this time, which is why it lives in the bayarea
directory.  It is meant to make it easier to access data from the
many different machines on which the model can run.  Invoke
'hopus -h' for more detail.

At this point, mounting and unmounting run directories is
supported.

git-svn-id: https://svn.urbansim.org/src/trunk@19153 1513fa5d-83ce-422e-ae47-6967bb0fb594