bayarea: add model to adjust dwelling units to household assignment
In the bayarea base year data, many buildings to which households
are assigned do not have enough dwelling units.  This simple
first-pass model cranks up the number of dwelling units to match
the number of households, and changes SF buildings to MF
buildings if necessary.  Naturally, it should only run in the
first model year, and should be the first model to run.

git-svn-id: https://svn.urbansim.org/src/trunk@19173 1513fa5d-83ce-422e-ae47-6967bb0fb594